Fork_Function
Fork Function is been used and a tree that is being made is shown in the image given there.
Fork Function is used to make a process to spilled into two processes you can call them a parent and child process which is helpful in the executing code to the efficient level.